Nam2/Ki2/KyThuatPhanMem/Lythuyet/BaiTapLonExample/2012-g1-ProjectFiles-1/Group#1_Work out with Friends/Code/1st version code/requirements.txt
Beaker==1.6.4
Chameleon==2.10
Jinja2==2.6
Mako==0.7.2
MarkupSafe==0.15
MySQL-python==1.2.4c1
PasteDeploy==1.5.0
Pillow==1.7.8
Pygments==1.5
SQLAlchemy==0.7.9
WTForms==1.0.2
WebHelpers==1.3
WebOb==1.2.3
WebTest==1.4.0
Whoosh==2.4.1
argparse==1.2.1
blueprint==3.4.0
coverage==3.6b1
distribute==0.6.32
mercurial==2.0.2
nose==1.2.1
py-bcrypt==0.2
py-pretty==1
pyramid==1.4a2
pyramid-beaker==0.7
pyramid-debugtoolbar==1.0.3
pyramid-exclog==0.6
pyramid-jinja2==1.4.2
pyramid-mailer==0.9
pyramid-rewrite==0.2
pyramid-tm==0.5
python-dateutil==2.1
pytz==2012h
repoze.lru==0.6
repoze.sendmail==3.2
six==1.2.0
times==0.6
timezones==1.5
transaction==1.3.0
translationstring==1.1
venusian==1.0a7
waitress==0.8.1
-e hg+https://code.google.com/p/workout-with-friends/@5f8a61e828a3b0fbd6d644ece7ba085e79fc1cac#egg=wowf-dev
wsgiref==0.1.2
zope.deprecation==4.0.0
zope.interface==4.0.1
zope.sqlalchemy==0.7.1
